requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bitarray==2.7.3
certifi==2023.5.7
charset-normalizer==3.1.0
cytoolz==0.12.1
decorator==5.1.1
Django==4.2.1
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-tester==0.9.0b1
eth-typing==3.3.0
eth-utils==2.1.0
executing==1.2.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
ipython==8.13.2
jedi==0.18.2
jsonschema==4.17.3
lru-dict==1.1.8
matplotlib-inline==0.1.6
multidict==6.0.4
parsimonious==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
protobuf==4.23.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.18.0
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.5
requests==2.30.0
rlp==3.0.0
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.2
toolz==0.12.0
traitlets==5.9.0
urllib3==2.0.2
wcwidth==0.2.6
web3==6.4.0
websockets==11.0.3
yarl==1.9.2
